Feat(Event): Add unregisterUser endpoint and implement logic in Event…

+++ src/main/java/com/alphasolutions/piauieventos/dto/UserRegistrationDTO.java
@@ -1,5 +1,5 @@
 package com.alphasolutions.piauieventos.dto;
 
 public record UserRegistrationDTO(
-        Long userId
+        Integer userId
 ) {}
